server/requirements.txt
streamlit==1.31.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.0
opencv-python-headless==4.9.0.80
numpy==1.26.0
Pillow==10.2.0
